[#2139] Refactoring of ReactiveStatelessDefaultBatchSizeTest
* Rename it to ReactiveStatelessWithBatchTest
* Clean up unit tests
